FA processing moved from preprocess_wgms_mbdata to preprocess_frontalablation
